+++ lib/recipe/bloc/recipe_event.dart
@@ -78,14 +78,90 @@ class ProbeChanged extends RecipeEvent {
   List<Object> get props => [recipe, probeName, probeIndex];
 }
 
+class ProbeChecked extends RecipeEvent {
+  final Recipe recipe;
+  final bool probeCheck;
+  final int probeIndex;
+
+  const ProbeChecked(
+      {required this.recipe,
+      required this.probeCheck,
+      required this.probeIndex});
+
+  @override
+  List<Object> get props => [recipe, probeCheck, probeIndex];
+}
+
 class ProbeDeleted extends RecipeEvent {
   final Recipe recipe;
-  final Map<String, dynamic> probe;
+  final Probe probe;
 
   const ProbeDeleted({required this.recipe, required this.probe});
 
   @override
   List<Object> get props => [recipe, probe];
+}
+
+class ProbeOptionAdded extends RecipeEvent {
+  final Recipe recipe;
+  final int probeIndex;
+
+  const ProbeOptionAdded({required this.recipe, required this.probeIndex});
+
+  @override
+  List<Object> get props => [recipe, probeIndex];
+}
+
+class ProbeOptionChanged extends RecipeEvent {
+  final Recipe recipe;
+  final int probeIndex;
+  final int probeOptionIndex;
+  final String probeOptionName;
+
+  const ProbeOptionChanged(
+      {required this.recipe,
+      required this.probeIndex,
+      required this.probeOptionIndex,
+      required this.probeOptionName});
+
+  @override
+  List<Object> get props =>
+      [recipe, probeIndex, probeOptionIndex, probeOptionName];
+}
+
+class ProbeOptionDeleted extends RecipeEvent {
+  final Recipe recipe;
+  final int probeIndex;
+  final int probeOptionIndex;
+
+  const ProbeOptionDeleted(
+      {required this.recipe,
+      required this.probeIndex,
+      required this.probeOptionIndex});
+
+  @override
+  List<Object> get props => [recipe, probeIndex, probeOptionIndex];
+}
+
+class ProbeOptionSelected extends RecipeEvent {
+  final Recipe recipe;
+  final int probeIndex;
+  final String answer;
+
+  const ProbeOptionSelected(
+      {required this.recipe, required this.probeIndex, required this.answer});
+
+  @override
+  List<Object> get props => [recipe, probeIndex, answer];
+}
+
+class RecipeProbesCleared extends RecipeEvent {
+  final Recipe recipe;
+
+  const RecipeProbesCleared({required this.recipe});
+
+  @override
+  List<Object> get props => [recipe];
 }
 
 class IngredientAdded extends RecipeEvent {