+++ lib/recipe/widgets/recipe_ingredients_screen.dart
@@ -3,13 +3,13 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gibsonify/recipe/recipe.dart';
 import 'package:gibsonify_api/gibsonify_api.dart';
 import 'package:gibsonify/navigation/navigation.dart';
-import 'package:gibsonify/collection/collection.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 
-class RecipeScreen extends StatelessWidget {
+class RecipeIngredientsScreen extends StatelessWidget {
   final int recipeIndex;
   final String? assignedFoodItemId;
-  const RecipeScreen(this.recipeIndex, {Key? key, this.assignedFoodItemId})
+  const RecipeIngredientsScreen(this.recipeIndex,
+      {Key? key, this.assignedFoodItemId})
       : super(key: key);
 
   @override
@@ -17,69 +17,27 @@ class RecipeScreen extends StatelessWidget {
     return BlocBuilder<RecipeBloc, RecipeState>(builder: (context, state) {
       return Scaffold(
           appBar: AppBar(
-            title: assignedFoodItemId == null
-                ? const Text('Add a new recipe')
-                : const Text('Confirm chosen recipe'),
+            title: const Text('Recipe ingredients'),
             leading: BackButton(
                 onPressed: () => {
                       context.read<RecipeBloc>().add(const RecipesSaved()),
                       Navigator.pop(context)
                     }),
-            actions: [
-              IconButton(
-                  onPressed: () => Navigator.pushNamed(
-                      context, PageRouter.recipeProbe,
-                      arguments: recipeIndex),
-                  icon: const Icon(Icons.help))
-            ],
           ),
           body: RecipeForm(recipeIndex),
-          floatingActionButton: Column(
-              mainAxisAlignment: MainAxisAlignment.end,
-              crossAxisAlignment: CrossAxisAlignment.end,
-              children: <Widget>[
-                FloatingActionButton.extended(
-                    heroTag: null,
-                    label: assignedFoodItemId == null
-                        ? const Text("Save Recipe")
-                        : const Text("Choose Recipe"),
-                    icon: assignedFoodItemId == null
-                        ? const Icon(Icons.save_sharp)
-                        : const Icon(Icons.check),
-                    onPressed: () {
-                      if (assignedFoodItemId == null) {
-                        context.read<RecipeBloc>().add(RecipeStatusChanged(
-                            recipe: state.recipes[recipeIndex],
-                            recipeSaved: true));
-                        context.read<RecipeBloc>().add(const RecipesSaved());
-                        Navigator.pop(context);
-                      } else {
-                        context.read<CollectionBloc>().add(
-                            FoodItemRecipeChanged(
-                                foodItemId: assignedFoodItemId!,
-                                foodItemRecipe: state.recipes[recipeIndex]));
-                        context.read<RecipeBloc>().add(const RecipesSaved());
-                        Navigator.pop(context);
-                        Navigator.pop(context);
-                      }
-                    }),
-                const SizedBox(
-                  height: 10,
-                ),
-                FloatingActionButton.extended(
-                    heroTag: null,
-                    label: const Text("New Ingredient"),
-                    icon: const Icon(Icons.add),
-                    onPressed: () => {
-                          context.read<RecipeBloc>().add(IngredientAdded(
-                              recipe: state.recipes[recipeIndex])),
-                          Navigator.pushNamed(context, PageRouter.ingredient,
-                              arguments: [
-                                recipeIndex,
-                                state.recipes[recipeIndex].ingredients.length
-                              ]),
-                        })
-              ]));
+          floatingActionButton: FloatingActionButton.extended(
+              heroTag: null,
+              label: const Text("New Ingredient"),
+              icon: const Icon(Icons.add),
+              onPressed: () => {
+                    context.read<RecipeBloc>().add(
+                        IngredientAdded(recipe: state.recipes[recipeIndex])),
+                    Navigator.pushNamed(context, PageRouter.ingredient,
+                        arguments: [
+                          recipeIndex,
+                          state.recipes[recipeIndex].ingredients.length
+                        ]),
+                  }));
     });
   }
 }
@@ -90,17 +48,22 @@ class RecipeForm extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Column(
-        children: <Widget>[
-          RecipeNameInput(recipeIndex),
-          RecipeNumberInput(recipeIndex),
-          RecipeVolumeInput(recipeIndex),
-          Ingredients(recipeIndex),
-        ],
-      ),
-    );
+    return BlocBuilder<RecipeBloc, RecipeState>(builder: (context, state) {
+      return Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Column(
+          children: <Widget>[
+            RecipeNameInput(recipeIndex),
+            const SizedBox(height: 10),
+            ListTile(
+                title: (state.recipes[recipeIndex].ingredients.isNotEmpty)
+                    ? const Text('Ingredients:')
+                    : null),
+            Ingredients(recipeIndex),
+          ],
+        ),
+      );
+    });
   }
 }
 
@@ -125,60 +88,6 @@ class RecipeNameInput extends StatelessWidget {
           onChanged: (value) {
             context.read<RecipeBloc>().add(RecipeNameChanged(
                 recipeName: value, recipe: state.recipes[recipeIndex]));
-          },
-          textInputAction: TextInputAction.next,
-        );
-      },
-    );
-  }
-}
-
-class RecipeNumberInput extends StatelessWidget {
-  final int recipeIndex;
-  const RecipeNumberInput(this.recipeIndex, {Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RecipeBloc, RecipeState>(
-      builder: (context, state) {
-        return TextFormField(
-          key: Key(state.recipes[recipeIndex].recipeNumber),
-          initialValue: state.recipes[recipeIndex].recipeNumber,
-          decoration: const InputDecoration(
-            icon: Icon(Icons.format_list_numbered),
-            labelText: 'Recipe Number',
-          ),
-          enabled: false,
-          textInputAction: TextInputAction.next,
-          keyboardType: TextInputType.number,
-        );
-      },
-    );
-  }
-}
-
-class RecipeVolumeInput extends StatelessWidget {
-  final int recipeIndex;
-  const RecipeVolumeInput(this.recipeIndex, {Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RecipeBloc, RecipeState>(
-      builder: (context, state) {
-        return TextFormField(
-          initialValue: state.recipes[recipeIndex].recipeVolume.value,
-          keyboardType: TextInputType.number,
-          decoration: InputDecoration(
-            icon: const Icon(Icons.play_for_work_rounded),
-            labelText: 'Recipe Volume',
-            helperText: 'Volume of recipe in ml e.g 250 ml',
-            errorText: state.recipes[recipeIndex].recipeVolume.invalid
-                ? 'Enter valid volume'
-                : null,
-          ),
-          onChanged: (value) {
-            context.read<RecipeBloc>().add(RecipeVolumeChanged(
-                recipeVolume: value, recipe: state.recipes[recipeIndex]));
           },
           textInputAction: TextInputAction.next,
         );